requirements.txt
Flask
psycopg2
pandas
scikit-learn
mysql-connector-python
